+++ packages/static-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/static-config
 
+## 3.2.0
+
+### Minor Changes
+
+- Support `maxDuration: 'max'` to allow the backend to resolve the maximum duration based on account plan type ([#15217](https://github.com/vercel/vercel/pull/15217))
+
 ## 3.1.2
 
 ### Patch Changes