+++ python/vercel-runtime/CHANGELOG.md
@@ -1,5 +1,17 @@
 # vercel-runtime
 
+## 0.8.0
+
+### Minor Changes
+
+- Fix error when running dev server on a django project. ([#15483](https://github.com/vercel/vercel/pull/15483))
+
+## 0.7.0
+
+### Minor Changes
+
+- [python] move vc_init_dev into vercel-runtime ([#15419](https://github.com/vercel/vercel/pull/15419))
+
 ## 0.6.0
 
 ### Minor Changes